#46: text_classifier.py, replace 'naive_bayes.py'
